app/requirements.txt
faust-streaming
kafka-python==2.0.2